Only allow dropping everything if env var set

+++ src/Database.ts
@@ -21,8 +21,13 @@ export default class Database<SCHEMA extends DatabaseSchema> {
 
 	/**
 	 * @deprecated WARNING: This drops and recreates your database!
+	 * 
+	 * Does nothing if `DEBUG_PG_ALLOW_DROP` is not set in your env.
 	 */
 	public async drop () {
+		if (!process.env.DEBUG_PG_ALLOW_DROP)
+			return;
+
 		return this.pool.query("DROP OWNED BY CURRENT_USER CASCADE");
 	}
 }
